Tidy up: moved plugins and generic validation rules in subfolders
