+++ tracker/core/tracker/src/Tracker.ts
@@ -3,10 +3,10 @@
  */
 
 import { AbstractGlobalContext, AbstractLocationContext, Contexts } from '@objectiv/schema';
-import { ApplicationContextPlugin } from './ApplicationContextPlugin';
 import { ContextsConfig } from './Context';
 import { waitForPromise } from './helpers';
-import { OpenTaxonomyValidationPlugin } from './OpenTaxonomyValidationPlugin';
+import { ApplicationContextPlugin } from './plugins/ApplicationContextPlugin';
+import { OpenTaxonomyValidationPlugin } from './plugins/OpenTaxonomyValidationPlugin';
 import { TrackerConsole } from './TrackerConsole';
 import { getLocationPath } from './TrackerElementLocations';
 import { TrackerEvent, TrackerEventConfig } from './TrackerEvent';
